requirements.txt
functions-framework==3.*
pandas
google-cloud-storage
flask
pyarrow
fastparquet